requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# Async & Workers
celery==5.3.4
aiohttp==3.9.1

# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3

# File Processing
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# Data Science
pandas==2.1.3
numpy==1.26.2

# NLP & AI
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
transformers==4.36.0
torch==2.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# APIs & Services
google-auth==2.23.4
google-api-python-client==2.108.0
msal==1.24.0
icalendar==5.0.10

# Cloud Storage
boto3==1.34.0
minio==7.2.2

# Monitoring
sentry-sdk==1.38.0

# Utilities
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2